fix(project): address comments from review
