+++ src/types/IRawInfer.ts
@@ -1,7 +1,7 @@
 import { IClique } from './IClique'
 import { ICliquePotentials } from './ICliquePotentials'
 
-export interface IrawInfer {
+export interface IRawInfer {
   cliques: IClique[];
   cliquesPotentials: ICliquePotentials;
 }
