+++ src/types/index.ts
@@ -21,4 +21,4 @@ export * from './INetworkResult'
 export * from './INodeResult'
 export * from './IInferAllOptions'
 export * from './IEvidence'
-export * from './IrawInfer'
+export * from './IRawInfer'
